Testing: Faster gitaly component upgrade
When Gitaly is not on or ahead of the version required, it will need
updating for the tests to be reliable. Currently this does remove the
Gitaly repository and child directories.

This is not great as a lot of work needs redoing:
1. Fetching the Git repository
2. Obtaining and rebuilding libgit2
3. Obtaining and rebuilding the Ruby gems (where not already present)

This process is time consuming and not required. The Gitaly Makefile is
rugged enough to understand what to rebuild when. As such this change
just tells the testing code that sets up Gitaly to keep the repository
around.

Side effects of this change are limited. Nonetheless note that the
inital clone is a shallow one, at `--depth=1`. This means that the
`git-fetch(1)` done by `GitLab::Taskhelpers#checkout_version` creates
a slightly more CPU intensive clone on the server side as the content
negotiation is more expensive.

+++ spec/support/helpers/test_env.rb
@@ -537,7 +537,7 @@ module TestEnv
     if component_needs_update?(install_dir, version)
       puts "\n==> Setting up #{component}..."
       # Cleanup the component entirely to ensure we start fresh
-      FileUtils.rm_rf(install_dir)
+      FileUtils.rm_rf(install_dir) unless component == "Gitaly"
 
       unless system('rake', task)
         raise ComponentFailedToInstallError
